refactor(l2): decouple eth_clients from SDK (#4083)
**Motivation**

Our two `EthClients` (L1/L2) and `SDK` are currently a mixture of
things.

- `EthClientL1` is polluted with L2 code.
- `EthClientL2` includes functions that could be placed in the `SDK`.


**Description**

This PR clarifies the responsibilities between these three components:

- `EthClientL1`: Limited to building and sending RPC requests.
- `EthClientL2`: Same as `EthClientL1`, but only contains L2 RPC
requests (`MessageProofs` and `Batches`).
- `SDK`: Contains more complex logic, such as `wait_for_receipt`,
`send_tx_bump`, etc.

As a result, we can now remove some duplicated code:
- `RpcBatch` was duplicated in both L1 and L2.
- `get_address_from_secret_key` was duplicated in both `EthClientL1` and
`SDK`, and is now placed under `l2::common::utils`. (It can't be placed
under the `SDK` because of circular dependencies with the `EthClientL2`.

The PR looks big, but it's mostly just moving things from one place to
another.
`GetBatchByNumberError` and `GetMessageProofError` haven't been moved to
L2 yet because it involves a bigger refactor, and I didn't want to make
the PR bigger.


Closes #3819
Closes #4007