chore(l1): added docker compose init + metric flags to snapsync make target (#4251)
**Motivation**

Add metrics + a docker compose with prometheus and grafana

**Description**

This PR add a couple of commands before the actual execution of the node
in the `server-sync` make target to start prometheus and grafana as
containers and later add the metrics flags to the node execution.

Locally this starts grafana at `3001` with default us/pass and allow to
see the block metrics as well as the resource (CPU/Memory/etc) ones.