fix(l2): remove empty account diffs from payload builder's blob size estimation (#4042)
**Motivation**

The payload builder estimates the state diff encoded size before adding
another transaction to the block to make sure that we can fit it in a
blob, and a batch can be committed with that single block.

When I executed load tests on the L2 the batch size metric reported 0,
so I initially suspected that the estimation was wrong (smaller than the
real state diff size) and I investigated the implementation but found
that the batch size metric was wrong instead, and the estimation was
correct*.

*Nevertheless I found that sometimes `get_account_diffs_in_tx()` returns
empty account diffs (I think because of accounts in the callframe's
original state that didn't participate in that transaction) and they get
accounted when estimating the state diff size, making the estimation
larger than the true value.

**Description**

- prevent `get_account_diffs_in_tx()` to retrieve empty account diffs
- add an internal error in the case that the sequencer can't make a
single block batch because of blob size, instead of getting stuck
silently.

Closes #issue_number

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>