feat(l1): chunk the download of headers (#4310)
**Motivation**

For large networks such as mainnet, downloading all the headers at once
causes high RAM usage since the headers are kept in memory until they're
all downloaded.

**Description**

We download them in batches (of 500k, should use ~300MB). Since the sync
head number is not known in the main sync loop, the cut is made in the
header fetching function.