feat(l1): add `--checkpoint` flag to `archive_sync` tool (#4231)
**Motivation**
Currenlty, we have no way of resuming an archive sync if it crashes/ we
need to stop it. Our only option is to begin from the start, which can
make us lose precious time in the case of unexpected crashes during long
archive syncs.
This PR proposes adding a `--checkpoint` flag which will periodically
write the current archive sync's checkpoint data (current root, last
hash, current file, etc) to a file. And which can later be used to
resume archive sync from that same checkpoint
<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: cdiielsi <49721261+cdiielsi@users.noreply.github.com>