feat(l1): block size limit (#4217)
**Description**

This PR implements [EIP-7934](https://eips.ethereum.org/EIPS/eip-7934),
which will be incorporated in Fusaka. Said EIP introduces a limit on the
size of the RLP encode of blocks. This PR adds the required check in
block validation.

Partially resolves #4158, the block building side of the EIP will be a
separate PR.