feat(l2): configure block max gas limit (#4211)
> [!CAUTION]
> Merge after #4224

**Motivation**

We need a way to limit the gas used per block to avoid overloading our
prover.

**Description**

- Adds a new CLI parameter: `max_gas_limit`.
- Updates the `payload_builder` to check this parameter as well.

**How to test**

Set a low `max_gas_limit`, start the L2, and watch in the monitor how
the initial deposits are distributed across different blocks, ensuring
that the gas used per block is lower than the configured amount.


Closes #4207 
Closes #2526