feat(replay): add `--cache-level` flag in `ethrex-replay` to decide whether to write the cache or not (#4402)
**Motivation**

This feature was already implemented in the `ethrex-replayer` but was
not available for `ethrex-replay`. As `ethrex-replayer` already makes
use of `ethrex-replay` as a library, it is almost free to add this
feature to `ethrex-replay` while keeping it in `ethrex-replayer`.

**Description**

- Rename `CacheLevel::All` to `CacheLevel::On` due to the change in the
context in which it is now used.
- Implements `--cache-level` arg of type `CacheLevel` with 3 variants:
`on` to cache block replays, `failed` to cache only failed replays, and
`off` to avoid caching.
- Refactors cache API
    - `write_cache` now derives the file name from the cache.
- Cache file name generation is now centralized in a single function to
reduce error surface.