feat(l1)!: change `--datadir` to be relative to current working directory (#4521)
May supersede https://github.com/lambdaclass/ethrex/pull/4514

**Motivation**

Currently, the datadir is specified relative to the application
directory (i.e. `/Users/<user>/Library/Application Support/` in macOS).
This behavior is confusing, since usually path args are relative to the
current working directory. Also, it seems to be prone to weird edge
cases like #4514.

**Description**

This PR changes this behavior by removing the path-prefixing behavior.
It also fixes the issue found in #4514, where any spaces in the path are
replaced by slashes (even if the path is the default path).