fix(l1): request storage peer leak (#4537)
**Motivation**

Whenever a step of snap sync is completed, the peers should be marked as
free. When requesting storage ranges, the function was ending without
clearing the peers if the block was stale.

**Description**

- Changes the request_storage_ranges function to wait until all spawned
tasks end and frees those peers before continuing.
- Adds utility function that frees the peers when they're supposed to be
free and logs an error if it finds peers in that inconsistent state.

This pr shouldn't be merged before thorough testing on Sepolia!