feat(l1): [EIP-7825] Set gas limit cap for tx and block validation (#4309)
**Motivation**

Implement EIP-7825 for Osaka fork.

**Description**

This pr introduces the limit cap of 16,777,216 gas (2^24) for individual
transactions, as indicated in the
[EIP-7825](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-7825.md).

This check is used in the `validate_transaction` and `validate_block`
methods of the blockchain crate and in `prepare_execution` for
transactions in `default_hooks`.

Also it updates the tests runners for running the tests associated to
the eip and it sets the gas_limit field of transactions for running the
state tests.

Closes #4154

---------

Co-authored-by: Jeremías Salomón 🐃🐄🥚 <48994069+JereSalo@users.noreply.github.com>
Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>