fix(l2): properly update message hashes in committer (#4547)
**Motivation**

We are prematurely updating the
[acc_messages](https://github.com/lambdaclass/ethrex/blob/101a11dce6e86332de330699a54dc549df03f651/crates/l2/sequencer/l1_committer.rs#L316)
variable in our committer, even though the block may later be discarded
due to `stateDiff` or privileged transaction limits.
The SP1 workflow is getting stuck in batches 1 or 2 for this reason. We
had not seen the error before because we weren’t spamming enough
transactions to hit the limit.
I also noticed that, since we are spamming 3000 privileged transactions
in our integration tests, 1000 retries to obtain a privileged receipt
falls short.

Successful run
[here](https://github.com/lambdaclass/ethrex/actions/runs/17838300648/job/50721020331).

**Description**

- Displays the sequencer logs in the SP1 workflow.  
- Increases the number of retries in `wait_for_transaction_receipt` to
10000.
- Correctly calculates the message hashes of the batch.  
- Re-enables the prover workflows on pushes to `main`.

Closes #4550