chore(l1): fix memory spike when transitioning from snap to fullsync (#4630)
**Motivation**

We were seeing spikes of more than 20GB of ram transitioning from
snapsync to full sync

**Description**

This PR remove the call to `into_fullsync` during the end of snap sync.
That call was unnecessary, `into_fullsync` is used if we try to start a
snap sync and we detect we are close to the head (<64 blocks) we move
the headers to the fullsync state in that case, but in this one we
already finished snapsync so we just need to start fullsync from the
last known head.