fix(replay): improve rate limit when using eth_getProofs (#4632)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
With the previous rate limit we were so slow that we couldn't even
execute replay with eth_getProof with a geth full node because it took
so long that the state was pruned (after 25min).
This new rate limit logic is a sweet spot, I had to make some changes to
original logic because it had some flaws.

Results: Execution can go from more than an hour to a few minutes
because of this