perf(l1): filter empty storages from snapshots (#4590)
During snap sync, specifically when downloading storages, we were
accidentally dumping all account addresses in every file, even when that
file had no storages for most of them.
By filtering, we reduce from 40GiB to 27GiB the auxiliary files in
Sepolia, and the sync goes from 3h10' to 3h2' in `ethrex-sync-3`, and
for mainnet in `ethrex-sync-2` we got from 185GiB to 40GiB and from 8h6'
to 7h7'.