fix(l1): preventing "Address family not supported by protocol" errors on discv4 (#4655)
**Motivation**

Using `UdpFramed` to send messages was causing a problem when a message
send failed: A singe error was causing every other message send to fail.
(See https://github.com/tokio-rs/tokio/issues/7648)
It was causing the log spam issue described in #4232 

**Description**

Moved back to use `UdpSocket` directly to send messages. It prevents the
"Address family not supported by protocol" propagation of errors.

Also fixes some of the issues in #4232 (but still full IPv6 needs to be
implemented):
- Comparing canonical IP addresses
- Removes a custom implementation of `to_canonical()`