feat(l1): use descriptive error when we fail to replace a mempool tx (#4656)
**Motivation**
We currently return a "NonceTooLow" error when we fail to replace a
mempool transaction (aka we receive a transaction with the same sender
and nonce as an existing one but without a higher fee value). This
doesn't make any sense and can be quite confusing. This PR adds a
MempoolError enum variant for this specific case.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Use a specific error variant when we fail to replace a mempool
transaction
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None