refactor(l1): spawnify kademlia (#4574)
**Motivation**

Current kademlia module is just a peer table that requires a lot of
Mutexes and locks() (and in fact it is not a kademlia implementation)

**Description**

This pr converts the PeerTable data structure into a GenServer in order
to avoid all the `Mutex.lock()` required in previous version and ensure
data consistency.

It is still a work in progress as I have to deal with the possible
errors when accessing PeerTable server.

Closes #4383